requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
cloudflare==2.11.1
idna==3.4
jsonlines==3.1.0
netifaces==0.11.0
PyYAML==6.0
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
